# perf: Optimize UTF8/ASCII byte offset index